Helix integration working and all the preliminary issues sorted out.
There are a bunch of undecided and untested things, but what is tested is the specialized flow for making reforge meshes from 2-dimensional splines. There is still a lot of work for axes creation and presentation for it to be truly effective.

Lots of refactors and work need to be done too to be honest going forward. But this unblocks a huge amount of the work.

+++ magicsim/magicsim/SimNode.cs
@@ -110,9 +110,9 @@ namespace magicsim
 
     public class Attribute
     {
-        public int strength;
-        public int agility;
-        public int intellect;
+        public double strength;
+        public double agility;
+        public double intellect;
     }
 
     public class Dps
